Choose a LICENSE
So I was about to go with MIT license . But then I was reading this 
https://opensource.stackexchange.com/questions/9805/can-i-license-my-project-with-an-open-source-license-but-disallow-commercial-use

And it made me realise, I think, that maybe what we want to do is go 'dual license' to start... cause we can always open up later. 

And I *think* that means Affero GPL v3. 

But I'd really like to hear people's thoughts on this one. Especially you @superarius  but actually everyone @karenkun , @jwu910 , @DAHuerta  included